backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
youtube-transcript-api==1.2.3
gunicorn==23.0.0
pytest==8.3.4
flake8==7.1.1
